% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropletMetadata.R
\docType{data}
\name{DropletMetadata}
\alias{DropletMetadata}
\title{Sample metadata from the Baran-Gale et al. ageing mouse thymus droplet experiments

A data frame containing mouse age and sort-type information for the
droplet single-cell RNA-sequencing experiment data sets}
\format{
A data frame containing information for each replicate sample

The object contains:
\describe{
\item{\code{sample}:}{Character, sample IDs.}
\item{\code{Gender}:}{Character, gender of mouse - all female.}
\item{\code{HTO}:}{Character, hashtag oligo assignments.}
\item{\code{ncells}:}{Integer, number of cells (post-QC) per sample.}
}
}
\usage{
DropletMetadata
}
\description{
Sample metadata from the Baran-Gale et al. ageing mouse thymus droplet experiments

A data frame containing mouse age and sort-type information for the
droplet single-cell RNA-sequencing experiment data sets
}
\examples{
head(DropletMetadata)

}
\references{
Baran-Gale J, Morgan MD, et al. (2020)
Ageing compromises mouse thymus function and remodels epithelial cell differentiation
\emph{eLife} 9:e56221.
}
\author{
Mike Morgan
}
\keyword{datasets}
