% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTseqMetadata.R
\docType{data}
\name{SMARTseqMetadata}
\alias{SMARTseqMetadata}
\title{Sample metadata from the Baran-Gale et al. ageing mouse thymus}
\format{
A data frame containing information for each replicate sample

The object contains:
\describe{
\item{\code{sample}:}{Character, day<1:5>.}
\item{\code{Age}:}{Character, age of mouse cells acquired from.}
\item{\code{Gender}:}{Character, gender of mouse - all female.}
\item{\code{ncells}:}{Integer, number of cells (post-QC) per sample.}
}
}
\usage{
SMARTseqMetadata
}
\description{
A data frame containing mouse age and sort-type information for the
SMARTseq experiment data sets
}
\examples{
head(SMARTseqMetadata)

}
\references{
Baran-Gale J, Morgan MD, et al. (2020)
Ageing compromises mouse thymus function and remodels epithelial cell differentiation
\emph{eLife} 9:e56221.
}
\author{
Mike Morgan
}
\keyword{datasets}
